cookiecutter-rag-template/{{cookiecutter.project_slug}}/requirements.txt
faiss-cpu
openinference
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation
openinference-instrumentation-langchain
opentelemetry-exporter-prometheus
